==> 1_Data_Collection.ipynb <==
# ASOIAF/GoT Reddit Posts - Pt. 1

## Data Collection

###### Import Libraries

import requests
import time
import pandas as pd
import json

###### Import Data

## creating a user-agent to access Reddit's content
headers = {'User-agent': 'JObot'}

## setting our subreddit links to loop through
subreddits = ['r/asoiaf/new/.json', 'r/gameofthrones/top/.json?t=all']

### Day 1 scraping

## setting empty posts list to collect our json 'data'
posts = []

## looping through the two different subreddits
for i in range(len(subreddits)):
    
    ## setting our after to None to start at the beginning of the thread
    after = None
    
    ## looping through 20 times and collecting 25 posts per loop
    for posts_25 in range(20):
        
        print(posts_25)  ## printing the computed iteration
        
        if after == None:  ## if statement to check value of 'after'
            params = {}    ## setting params to empty dictionary if 'None'
        
        else:
            params = {'after': after}  ## setting our params to the after at the end of each series of posts
        
        ## creating the url for given subreddit by attaching the looped through subreddit
        url = 'https://www.reddit.com/' + subreddits[i]  
        
        ## using our get request based on url and current params
        res = requests.get(url, params = params, headers=headers)
        
        ## if statement run, if our status code is 200 (good)
        if res.status_code == 200:
            asoiaf = res.json() ## gathering/assigning json info from our get request

            ## extending our list of posts from data in 'children'
            posts.extend(asoiaf['data']['children'])

            ## setting our after value to
            after = asoiaf['data']['after']

        ## if error with status code, we will break our loop and print error code
        else:
            print(res.status_code)  
            break

        time.sleep(1) ## slowing our loop by one second for each iteration

## showing that we have 1004 unique posts
len(set([p['data']['name'] for p in posts]))

## looking at our keys, so that we can isolate our text from the post
posts[0]['data'].keys()

## showing our first post in posts
posts[0]['data']

subreddit_lists = []
for post in range(len(posts)):
    post_dict = {}
    post_dict['subreddit'] = posts[post]['data']['subreddit']
    post_dict['post_text'] = posts[post]['data']['selftext']
    
    if post_dict['post_text'] != '':
        subreddit_lists.append(post_dict)
    else:
        pass

subreddit_lists[-1]

df = pd.DataFrame(subreddit_lists)

df.shape

df.info()

df.to_csv('day_1_reddit_scrapping.csv', index=False)

df.to_csv('day_1_reddit_scraping_indexed.csv')

### Day 2 scraping

posts = []
afters = []
for i in range(len(subreddits)):
    after = None
    for posts_25 in range(25):
        print(posts_25)
        if after == None:
            params = {}
        else:
            params = {'after': after}
        url = 'https://www.reddit.com/' + subreddits[i]
        res = requests.get(url, params = params, headers=headers)
        if res.status_code == 200:
            asoiaf = res.json()
            
            ## append will do a list of list, but extend will add to the list
            posts.extend(asoiaf['data']['children'])

            ## this is where we overwrite it, we start at zero on the first hit
            ## then we will set it to a value after
            ## now we can hit this API with this 'after' paramater and get 25 more hits every time
            after = asoiaf['data']['after']

        else:
            print(res.status_code)
            break

        ## slow your loop down intentionally, so you're not hitting the server hard
        time.sleep(1) ## slows the loop down by one second for each iteration
        afters.append(after) ## collecting our after values to start each successive loop with

len(posts)

recent_afters = [afters[24], afters[49]]

recent_afters

def sub_collection(subreddit_list):
    posts = []
    for i in range(len(subreddit_list)):
        after = None
        for posts_25 in range(40):
            if after == None:
                params = {}
            else:
                params = {'after': after}
            url = 'https://www.reddit.com/' + subreddit_list[i] 
            res = requests.get(url, params = params, headers=headers)
            if res.status_code == 200:
                asoiaf = res.json()

                ## append will do a list of list, but extend will add to the list
                posts.extend(asoiaf['data']['children'])

                ## this is where we overwrite it, we start at zero on the first hit
                ## then we will set it to a value after
                ## now we can hit this API with this 'after' paramater and get 25 more hits every time
                after = asoiaf['data']['after']

            else:
                print(res.status_code)
                break

            ## slow your loop down intentionally, so you're not hitting the server hard
            time.sleep(2) ## slows the loop down by two seconds for each iteration            
    
     ## creating a function that will separate the text and the respective subreddit
#     def text_parsing(posts):
    subreddit_lists = []
    for post in range(len(posts)):
        post_dict = {}
        post_dict['subreddit'] = posts[post]['data']['subreddit']
        post_dict['post_text'] = posts[post]['data']['selftext']

        if post_dict['post_text'] != '':
            subreddit_lists.append(post_dict)
        else:
            pass

    return subreddit_lists

new_posts, new_afters = sub_collection(subreddits, recent_afters)

## afters for #3
new_afters

## creating a function that will separate the text and the respective subreddit
def text_parsing(posts):
    subreddit_lists = []
    for post in range(len(posts)):
        post_dict = {}
        post_dict['subreddit'] = posts[post]['data']['subreddit']
        post_dict['post_text'] = posts[post]['data']['selftext']

        if post_dict['post_text'] != '':
            subreddit_lists.append(post_dict)
        else:
            pass
    return subreddit_lists

posts_2 = text_parsing(posts)

## showing the length of collected posts
len(posts_2)

df2 = pd.DataFrame(posts_2)
df2.head()

df2.to_csv('day_2_reddit_scrapping.csv')

### Day 3 Scraping

posts_3 = sub_collection(subreddits)

len(posts_3)

posts_3[-1]

df3 = pd.DataFrame(posts_3)
df3.head()

df3.subreddit.value_counts()

df3.to_csv('day_3_reddit_scraping.csv')

### Day 4 scraping

posts_4 = sub_collection(subreddits)
len(posts_4)

df4 = pd.DataFrame(posts_4)
df4.head()

df4.subreddit.value_counts()

df4.to_csv('day_4_reddit_scraping.csv')

### Day 5 scraping

## changed the subreddit list to include all new posts (no long 'hot')

posts_5 = sub_collection(subreddits)
len(posts_5)

df5 = pd.DataFrame(posts_5)
df5.head()

df5.subreddit.value_counts()

df5.to_csv('day_5_reddit_scraping.csv')

### Day 6 scraping

posts_6 = sub_collection(subreddits)
len(posts_6)

df6 = pd.DataFrame(posts_6)
df6.head()

df6.subreddit.value_counts()

df6.to_csv('day_6_reddit_scraping.csv')

## scraping the hots section of 
posts_6 = sub_collection(subreddits)
len(posts_6)

df6_hot = pd.DataFrame(posts_6)
df6_hot.head()

df6 = pd.concat([df6, df6_hot], ignore_index=True)
df6.tail()

## scraping the top for the month
posts_6t = sub_collection(subreddits)
len(posts_6t)

## scraping the controversial for the past month
posts_6c = sub_collection(subreddits)
len(posts_6c)

df6t = pd.DataFrame(posts_6t)
df6c = pd.DataFrame(posts_6c)

df6 = pd.concat([df6, df6_hot, df6t, df6c], ignore_index=True)
df6.shape

df6.to_csv('day_6_reddit_scraping.csv')

### Day 7 scraping

## pulling our final days worth of subreddits
posts_7 = sub_collection(subreddits)
df7 = pd.DataFrame(posts_7)
print(len(posts_7))

df7.subreddit.value_counts()

no_asoiaf = df7[df7['subreddit'] == 'gameofthrones']
no_asoiaf.subreddit.value_counts()

no_asoiaf.to_csv('day_7_reddit_scraping.csv')
