==> 2_EDA_and_Preprocessing.ipynb <==
# ASOIAF/GoT Reddit Posts - Pt. 2

># EDA

### Import Libraries

import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np

%matplotlib inline

### Load Data

## loading in all of our data
day_1 = pd.read_csv('./data/day_1_reddit_scraping_indexed.csv')
day_2 = pd.read_csv('./data/day_2_reddit_scrapping.csv')
day_3 = pd.read_csv('./data/day_3_reddit_scraping.csv')
day_4 = pd.read_csv('./data/day_4_reddit_scraping.csv')
day_5 = pd.read_csv('./data/day_5_reddit_scraping.csv')
day_6 = pd.read_csv('./data/day_6_reddit_scraping.csv')
day_7 = pd.read_csv('./data/day_7_reddit_scraping.csv')

## concatenating all our datasets
data = pd.concat([day_1, day_2, day_3, day_4, day_5, day_6, day_7], ignore_index=True)

## viewing the shape of our data, we'll need to drop our indexed column
data.shape

## looking at our initial data
data.head()

## getting rid of the "unnamed" column
data.drop("Unnamed: 0", axis=1, inplace = True)

## locating the duplicate text in our posts
data = data.loc[pd.DataFrame.duplicated(data) == False]

## verifying that there is no more duplicates
pd.DataFrame.duplicated(data).value_counts()

## showing corrected header
data.head()

## checking our value counts
data.subreddit.value_counts()

### Pre-processing

## no null values
data.isnull().sum().sum()

## viewing the shape of our data
data.shape

data.describe()

### Attempt at Visualizing with Spacy / Scattertext

import scattertext as st
import re, io
from pprint import pprint
from scipy.stats import rankdata, hmean, norm
import spacy
import os, pkgutil, json, urllib
from urllib.request import urlopen
from IPython.display import IFrame
from IPython.core.display import display, HTML
display(HTML("&lt;style>.container { width:98% !important; }&lt;/style>"))

from spacy.lang.en import English
from spacy.lang.en.stop_words import STOP_WORDS

from scattertext import CorpusFromPandas, produce_scattertext_explorer

## loading in spaCy
nlp = spacy.load('en_core_web_sm')

## creating our parsed/tokenized text
data['parsed_post'] = data['post_text'].apply(nlp)

## showing new column
data.head()

## loading spacy stop_words
spacy_stopwords = spacy.lang.en.stop_words.STOP_WORDS
print('Number of stop words: %d' % len(spacy_stopwords))
print('First ten stop words: %s' % list(spacy_stopwords)[:10])

## creating a corpus to use for our scattertext plot
corpus = st.CorpusFromPandas(data, category_col='subreddit', text_col='post_text', nlp=nlp).build()

## The most frequent terms in our corpus
list(corpus.get_scaled_f_scores_vs_background().index[:10])

## turning our scattertext corpus into a dataframe
term_freq_df = corpus.get_term_freq_df() 

## creating scaled scores for 'asoiaf'
term_freq_df['asoiaf_scaled'] = corpus.get_scaled_f_scores(category='asoiaf') 

## creating scaled scores for 'got'
term_freq_df['got_scaled'] = corpus.get_scaled_f_scores(category='gameofthrones') 

## showing the top scaled 'asoiaf' values
term_freq_df.sort_values(by='asoiaf_scaled', ascending=False).head(10)

## creating a column of scaled value differences
term_freq_df['scaled_diff'] = abs(term_freq_df["asoiaf_scaled"].add(-term_freq_df["got_scaled"]))

## separating out our values that occur frequently in both subreddit
stop_phrases = term_freq_df[term_freq_df['scaled_diff'] <= 0.1]
stop_phrases.shape

stop_phrases.head()

## making the dataframe only the index (i.e. just the words)
stop_phrases2 = stop_phrases.drop(labels=['asoiaf freq', 
                                          'gameofthrones freq', 
                                          'asoiaf_scaled', 
                                          'got_scaled', 
                                          'scaled_diff'], 
                                 axis=1)
stop_phrases2.head()

## setting our list of stop_phrases to a csv, so we can turn them into a list
## we will later use this list in our model hypertuning
stop_phrases2.to_csv('stop_list.csv', index=True) 

## creating an html 'scattertext' based on scaled data
## code is from scattertext's creator, Jason Kessler
html = produce_scattertext_explorer(corpus,
                                    category='asoiaf',
                                    category_name='ASOIAF',
                                    not_category_name='Game of Thrones',
                                    width_in_pixels=1200,
                                    show_characteristic = False,
                                    minimum_term_frequency=5,
                                   )
file_name = './data/ASOIAF_GoT_Scattertext_Scale.html'  ## saving html file
open(file_name, 'wb').write(html.encode('utf-8'))  ## pulling file from scattertext library
IFrame(src=file_name, width = 1400, height=1000)  ## using Iframe to display our file

## creating an html 'scattertext' based on scaled data
## code is from scattertext's creator, Jason Kessler
html = produce_scattertext_explorer(corpus,
                                    category='asoiaf',
                                    category_name='ASOIAF',
                                    not_category_name='Game of Thrones',
                                    width_in_pixels=1200,
                                    minimum_term_frequency=5,
                                    show_characteristic = False,
                                    transform=st.Scalers.log_scale_standardize
                                   )
file_name = './data/ASOIAF_GoT_Scattertext_Log.html'
open(file_name, 'wb').write(html.encode('utf-8'))
IFrame(src=file_name, width = 1400, height=1000)

###### Export to CSV

data.to_csv('unique_data.csv', index=False)
