==> 3_Feature_Engineering_and_Baseline.ipynb <==
# ASOIAF/GoT Reddit Posts - Pt. 3

># Feature Engineering and Baseline Model 

### Import Libraries

from spacy.lang.en import English
from spacy.lang.en.stop_words import STOP_WORDS
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
import spacy
from spacy.util import minibatch, compounding
from spacy.pipeline import TextCategorizer


%matplotlib inline

### Reload Cleaned Data

## loading our data
data = pd.read_csv('./data/unique_data.csv')

## showing first 5 rows
data.head()

## binarizing our subreddit category
data['subreddit'] = data['subreddit'].map({'asoiaf':0,
                               'gameofthrones':1})
## showing success
data['subreddit'].value_counts()

## importing our custom stop words
stop_list = pd.read_csv('./data/stop_list.csv')
stop_list.head()

## creating our stop word/phrase list
custom_stop = []
for word in range(len(stop_list['term'])):
    custom_stop.append(stop_list['term'][word])

## printing our first 5
custom_stop[:5]

### Baseline Model

## setting our X and target(y) variables
X = data['parsed_post']
y = data['subreddit']

## showing our baseline percentage
y.value_counts(normalize=True)

>## Building a Model with spaCy

## loading spaCy
nlp = spacy.load('en_core_web_sm')

## using a lambda function to turn all parsed_posts and subreddits into a Series containing tuples
data['tuples'] = data.apply(
    lambda row: (row['parsed_post'],row['subreddit']), axis=1)

## turning the tuples into a list we'll use for training
train = data['tuples'].tolist()

## showing our first tuple
train[0]

## SOURCE: functions mainly manufactured from spacy documentation

## creating a function that will load are training size and explicitly call our limit
def load_data(limit=0, train_size=0.8):
    
    train_data = train  ## setting our training data
    np.random.shuffle(train_data)  ## randomly shuffling our training data
    
    ## setting the limit based on function being passed
    ## with current code it will be set to n_texts value below
    train_data = train_data[-limit:]
    
    ## unzipping our tuple into two separate lists (texts and subreddit value/type)
    texts, bool_val = zip(*train_data)
    
    ## setting our subreddit values to True or False
    ## this will be used in the scoring portion of the 'evaluate' function
    cats = [{'Game of Thrones': bool(y)} for y in bool_val]
    split = int(len(train_data) * train_size)  ## splits our training data based on argument passed
    
    ## returns our text and category (True/False vals) test/train split
    return (texts[:split], cats[:split]), (texts[split:], cats[split:])



## creating a function that will score our data after instantiating our optimizer
def evaluate(tokenizer, textcat, texts, cats):
    
    ## setting our docs for 
    docs = (tokenizer(text) for text in texts)
    
    ## setting empty values for our confusion matrix and scores
    tp = 1e-8 # True positives , 1e-8 is there to ensure we never divide by zero
    fp = 0 # False positives
    fn = 0  # False negatives
    tn = 0  # True negatives
    
    ## using our textcat.pipe on our tokenized docs to pull a number and it's associated value
    for i, doc in enumerate(textcat.pipe(docs)):
        
        gold = cats[i] ## setting our 'gold_parse' to the indexed dict value of {'GoT': T or F} 
        
        ## for_loop to extract boolean values of 'gold_parse' and doc 'score' from textcat.pipe
        for bool_val, score in doc.cats.items():
            
            ## a series of if_statements that will increase count on our confusion matrix
            ## when the each parameter for 'score' and 'bool_val' are met
            if bool_val not in gold:
                continue
            if score >= 0.5 and gold[bool_val] >= 0.5:
                tp += 1.
            elif score >= 0.5 and gold[bool_val] < 0.5:
                fp += 1.
            elif score < 0.5 and gold[bool_val] < 0.5:
                tn += 1
            elif score < 0.5 and gold[bool_val] >= 0.5:
                fn += 1
    
    ## calculate and return the following scores
    accuracy = (tp + tn) / (tp + tn + fp + fn)            
    precision = tp / (tp + fp)
    recall = tp / (tp + fn)
    f_score = 2 * (precision * recall) / (precision + recall)
    
    return {'textcat_p': precision, 
            'textcat_r': recall, 
            'textcat_f': f_score, 
            'textcat_a': accuracy}

## number of texts to train
n_texts=len(data)

## number of training iterations
n_iter=20

## this is the protocol for calling spaCy's TextCategorizer'textcat'
## Code adapted from spaCy's documentation to create or add the necessary pipeline
if 'textcat' not in nlp.pipe_names:
    textcat = nlp.create_pipe('textcat')
    nlp.add_pipe(textcat, last=True)
# otherwise, get it, so we can add labels to it
else:
    textcat = nlp.get_pipe('textcat')

## adding label to text classifier, used for Boolean comparison
textcat.add_label('Game of Thrones')

## this creates our train/test split from the 'load_data' function
(train_texts, train_cats), (test_texts, test_cats) = load_data(limit=n_texts)  

## prints details of line above, samples size and split
print(f'Using {n_texts} examples ({len(train_texts)} training, {len(test_texts)} evaluation)')

## takes 'train_text' and 'bool_val' from load_data function
## this will get evaluated in the next block of code
train_data = list(zip(train_texts,
                      [{'cats': cats} for cats in train_cats]))

## Some more code adapted from spaCy
## this block ensures that we are only calling and training one pipeline 'textcat'
other_pipes = [pipe for pipe in nlp.pipe_names if pipe != 'textcat']
with nlp.disable_pipes(*other_pipes):  # only train textcat
    
    ## spaCy trains a model with begin_training() using thinc.neural
    optimizer = nlp.begin_training()
    print("Training the model...")
    
    ## redo as an f-string // Header column
    print(f'{"LOSS":^8}{"P":^8}{"R":^8}{"F":^8}{"A":^8}')
    
    for i in range(n_iter):  ## computing our set number of iterations
        losses = {}  ## creating a dictionary to store our losses
        
        ## using spaCy's minibatching to create a smaller batches of our data 
        ## for gradient descent optimization
        batches = minibatch(train_data, size=compounding(4., 32., 1.001))
        
        ## looping through the smaller batches
        for batch in batches:
            ## unzipping our batches and assigning the post text to 'texts' 
            ## and the 'cats' and 'bool_val' to 'annotations'
            texts, annotations = zip(*batch)
            
            ## updating our model by using our 'optimizer' for gradient descent
            ## setting a dropout rate of .2 to prevent overfitting our model
            ## reevaluating our losses
            nlp.update(texts, annotations, sgd=optimizer, drop=0.2,
                       losses=losses)
        
          ## encoding the parameters of thinc.neural on 'textcat'
        with textcat.model.use_params(optimizer.averages):
            
            # evaluate on the dev data split off in load_data()
            ## evaluate is the function from above
            scores = evaluate(nlp.tokenizer, textcat, test_texts, test_cats)
          
        ## printing results (rounded, with centering and 8 spaces)
        print(f'{round(losses["textcat"],2):^8}'  
              f'{round(scores["textcat_p"],3):^8}' 
              f'{round(scores["textcat_r"],3):^8}'
              f'{round(scores["textcat_f"],3):^8}'
              f'{round(scores["textcat_a"],3):^8}'
             )


>## Modeling via Sci-kit Learn

from sklearn.model_selection import GridSearchCV, train_test_split
from sklearn.naive_bayes import MultinomialNB, BernoulliNB
from sklearn.linear_model import LogisticRegression
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.pipeline import Pipeline
from sklearn.metrics import accuracy_score, scorer
from sklearn.svm import SVC
from sklearn.ensemble import RandomForestClassifier, VotingClassifier
import sklearn

## reminder of our baseline
y.value_counts(normalize=True)

## creating our train/test split
X_train, X_test, y_train, y_test = train_test_split(X, y, 
                                                    test_size=.3, 
                                                    random_state=42, 
                                                    stratify=y)

## creating a list of our two chosen vectorizers to iterate through in our grid search
vectorizer = [CountVectorizer(), TfidfVectorizer()]

## creating variables to accept tuning parameters
max_feat = [300, 500]  
ngram_range = [(1, 3), (1, 2)] 
stop_words = [None, 'english', custom_stop] 
max_df = [0.9, 0.8] 

## creating any empty results list to capture our cv_results_ at the end of each iteration
results = []

## looping through both vectorizers
for vect in vectorizer:
    
    ## creating a pipeline for our vectorizer and classifier models
    pipeline = Pipeline([
        ('vect', vect),
        ('clf', LogisticRegression())
    ])
    
    ## generating our parameters for vectorizers
    vect_params = {'vect__max_features': max_feat,
                    'vect__stop_words': stop_words,
                    'vect__ngram_range': ngram_range,
                    'vect__max_df': max_df
                  }
    parameters = [
        {
            ## Logistic Regression
            'vect__max_features': max_feat,
            'vect__stop_words': stop_words,
            'vect__ngram_range': ngram_range,
            'vect__max_df': max_df,
            'clf': (LogisticRegression(solver='liblinear'), ), ## setting our first classifier model
            'clf__penalty': ('l1', 'l2'), #2
            'clf__C': (.5, 1.0), #4 288*2
        }, 
        {
            ## Multinomial Bayes
            'vect__max_features': max_feat,
            'vect__stop_words': stop_words,
            'vect__ngram_range': ngram_range,
            'vect__max_df': max_df,
            'clf': (MultinomialNB(), ),  ## setting our second classifier model
            'clf__alpha': (.5, 1.0)  #2 72*2
        },
        {
            ## SVC
            'vect__max_features': max_feat,
            'vect__stop_words': stop_words,
            'vect__ngram_range': ngram_range,
            'vect__max_df': max_df,
            'clf': (SVC(gamma='scale', ), ),
            'clf__kernel': ('rbf', 'poly') 
        },
        {
            ## RandomForestClassifier
            'vect__max_features': max_feat,
            'vect__stop_words': stop_words,
            'vect__ngram_range': ngram_range,
            'vect__max_df': max_df,
            'clf': (RandomForestClassifier(n_estimators=50, min_samples_split=5), ),
        },
        {
            ## putting together an ensemble model
            'vect__max_features': max_feat,
            'vect__stop_words': stop_words,
            'vect__ngram_range': ngram_range,
            'vect__max_df': max_df,
            'clf': (VotingClassifier(estimators=[('lr', LogisticRegression()), 
                                                 ('rf', RandomForestClassifier()), 
                                                 ('mnb', MultinomialNB()), 
                                                 ('svc', SVC())],                                           
                                            voting='hard'), )
        }
        
    ]
    
    ## performing our grid search with the inherited pipeline and parameters
    grid_search = GridSearchCV(pipeline, 
                               parameters,
                               cv=3,
                               n_jobs=-1,
                               verbose=1,
                               return_train_score=True
                              )
    
    ## running an if statement to print the type of vectorizer used
    if vect == vectorizer[0]:
        vect_string = "CountVectorizer"
    
    else:
        vect_string = "Tf-IDF Vectorizer"
    
    ## fitting our model and printing our best scores and parameters
    grid_search.fit(X_train, y_train)
    print(f'''Best score for {vect_string} is: 
    {round(grid_search.best_score_, 4)}
    ''')
    print(grid_search.best_params_)
    print("")
    
    ## appending our cv_results_ to the end of results
    results.append(grid_search.cv_results_)

## turning our results into a CV
results_df = pd.DataFrame(grid_search.cv_results_)

## printing out our results
results_df
